requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
curl_cffi==0.7.0b4
distro==1.9.0
duckduckgo_search==5.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
openai==1.29.0
orjson==3.10.3
phidata==2.4.7
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
tomli==2.0.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
